requirements.txt
googlemaps==4.10.0
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
Pillow==10.0.0
flask-restful==0.3.10
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
flask-cors==5.0.0
